Better stroke implementation, as a double queue (relies on LinkedList, but with thread synchronization).
Emptiness now uses a nicer law: A*ln(1/(F* x+1))+127 that is 127 when x, the number of points is 0, and is 0 when x is the maximum number of points.
A must be calculated as:  A = -(127 / ln(1/(F*M+1))), where M is the maximum number of points (x size x y size). F is a scaling factor, to make it more reactive at the beginning.